(#43) - remove TypedArray shim
I'm abandoning #44 in favor of this
PR.  It's better to stop lying to
ourselves about which browsers we
support by shimming it in the tests.

Plus, shimming it in the tests is
a good way to hide bugs, and the
npm typedarray package would
be better than rolling our own.